factor out CI defaults
