+++ helpers/providers/index.ts
@@ -1,4 +1,3 @@
-import ciInfo from "ci-info";
 import prompts from "prompts";
 import { questionHandlers } from "../../questions/utils";
 import { ModelConfig, ModelProvider, TemplateFramework } from "../types";
@@ -27,7 +26,7 @@ export async function askModelConfig({
   framework,
 }: ModelConfigQuestionsParams): Promise<ModelConfig> {
   let modelProvider: ModelProvider = DEFAULT_MODEL_PROVIDER;
-  if (askModels && !ciInfo.isCI) {
+  if (askModels) {
     let choices = [
       { title: "OpenAI", value: "openai" },
       { title: "Groq", value: "groq" },
