+++ questions/questions.ts
@@ -1,4 +1,3 @@
-import ciInfo from "ci-info";
 import { blue, green } from "picocolors";
 import prompts from "prompts";
 import { COMMUNITY_OWNER, COMMUNITY_REPO } from "../helpers/constant";
@@ -17,71 +16,36 @@ import {
   selectLocalContextData,
 } from "./utils";
 
-const defaults: Omit<QuestionArgs, "modelConfig"> = {
-  template: "streaming",
-  framework: "nextjs",
-  ui: "shadcn",
-  frontend: false,
-  llamaCloudKey: "",
-  useLlamaParse: false,
-  communityProjectConfig: undefined,
-  llamapack: "",
-  postInstallAction: "dependencies",
-  dataSources: [],
-  tools: [],
-};
-
 export const askProQuestions = async (program: QuestionArgs) => {
-  const getDefault = <K extends keyof Omit<QuestionArgs, "modelConfig">>(
-    field: K,
-  ): Omit<QuestionArgs, "modelConfig">[K] => defaults[field];
-
-  // Ask for next action after installation
-  async function checkAskPostInstallAction() {
-    if (program.postInstallAction === undefined) {
-      if (ciInfo.isCI) {
-        program.postInstallAction = getDefault("postInstallAction");
-      } else {
-        program.postInstallAction = await askPostInstallAction(program);
-      }
-    }
-  }
-
-  if (!program.template) {
-    if (ciInfo.isCI) {
-      program.template = getDefault("template");
-    } else {
-      const styledRepo = blue(
-        `https://github.com/${COMMUNITY_OWNER}/${COMMUNITY_REPO}`,
-      );
-      const { template } = await prompts(
+  const styledRepo = blue(
+    `https://github.com/${COMMUNITY_OWNER}/${COMMUNITY_REPO}`,
+  );
+  const { template } = await prompts(
+    {
+      type: "select",
+      name: "template",
+      message: "Which template would you like to use?",
+      choices: [
+        { title: "Agentic RAG (e.g. chat with docs)", value: "streaming" },
         {
-          type: "select",
-          name: "template",
-          message: "Which template would you like to use?",
-          choices: [
-            { title: "Agentic RAG (e.g. chat with docs)", value: "streaming" },
-            {
-              title: "Multi-agent app (using workflows)",
-              value: "multiagent",
-            },
-            { title: "Structured Extractor", value: "extractor" },
-            {
-              title: `Community template from ${styledRepo}`,
-              value: "community",
-            },
-            {
-              title: "Example using a LlamaPack",
-              value: "llamapack",
-            },
-          ],
-          initial: 0,
+          title: "Multi-agent app (using workflows)",
+          value: "multiagent",
         },
-        questionHandlers,
-      );
-      program.template = template;
-    }
-  }
+        { title: "Structured Extractor", value: "extractor" },
+        {
+          title: `Community template from ${styledRepo}`,
+          value: "community",
+        },
+        {
+          title: "Example using a LlamaPack",
+          value: "llamapack",
+        },
+      ],
+      initial: 0,
+    },
+    questionHandlers,
+  );
+  program.template = template;
 
   if (program.template === "community") {
     const projectOptions = await getProjectOptions(
@@ -122,7 +86,7 @@ export const askProQuestions = async (program: QuestionArgs) => {
       questionHandlers,
     );
     program.llamapack = llamapack;
-    await checkAskPostInstallAction();
+    program.postInstallAction = await askPostInstallAction(program);
     return; // early return - no further questions needed for llamapack projects
   }
 
@@ -132,92 +96,76 @@ export const askProQuestions = async (program: QuestionArgs) => {
     program.dataSources = [EXAMPLE_FILE];
     program.framework = "fastapi";
   }
-  if (!program.framework) {
-    if (ciInfo.isCI) {
-      program.framework = getDefault("framework");
-    } else {
-      const choices = [
-        { title: "NextJS", value: "nextjs" },
-        { title: "Express", value: "express" },
-        { title: "FastAPI (Python)", value: "fastapi" },
-      ];
+  const choices = [
+    { title: "NextJS", value: "nextjs" },
+    { title: "Express", value: "express" },
+    { title: "FastAPI (Python)", value: "fastapi" },
+  ];
 
-      const { framework } = await prompts(
-        {
-          type: "select",
-          name: "framework",
-          message: "Which framework would you like to use?",
-          choices,
-          initial: 0,
-        },
-        questionHandlers,
-      );
-      program.framework = framework;
-    }
-  }
+  const { framework } = await prompts(
+    {
+      type: "select",
+      name: "framework",
+      message: "Which framework would you like to use?",
+      choices,
+      initial: 0,
+    },
+    questionHandlers,
+  );
+  program.framework = framework;
 
   if (
     (program.framework === "express" || program.framework === "fastapi") &&
     (program.template === "streaming" || program.template === "multiagent")
   ) {
     // if a backend-only framework is selected, ask whether we should create a frontend
-    if (program.frontend === undefined) {
-      if (ciInfo.isCI) {
-        program.frontend = getDefault("frontend");
-      } else {
-        const styledNextJS = blue("NextJS");
-        const styledBackend = green(
-          program.framework === "express"
-            ? "Express "
-            : program.framework === "fastapi"
-              ? "FastAPI (Python) "
-              : "",
-        );
-        const { frontend } = await prompts({
-          onState: onPromptState,
-          type: "toggle",
-          name: "frontend",
-          message: `Would you like to generate a ${styledNextJS} frontend for your ${styledBackend}backend?`,
-          initial: getDefault("frontend"),
-          active: "Yes",
-          inactive: "No",
-        });
-        program.frontend = Boolean(frontend);
-      }
-    }
+    const styledNextJS = blue("NextJS");
+    const styledBackend = green(
+      program.framework === "express"
+        ? "Express "
+        : program.framework === "fastapi"
+          ? "FastAPI (Python) "
+          : "",
+    );
+    const { frontend } = await prompts({
+      onState: onPromptState,
+      type: "toggle",
+      name: "frontend",
+      message: `Would you like to generate a ${styledNextJS} frontend for your ${styledBackend}backend?`,
+      initial: false,
+      active: "Yes",
+      inactive: "No",
+    });
+    program.frontend = Boolean(frontend);
   } else {
     program.frontend = false;
   }
 
   if (program.framework === "nextjs" || program.frontend) {
     if (!program.ui) {
-      program.ui = defaults.ui;
+      program.ui = "shadcn";
     }
   }
 
   if (!program.observability && program.template === "streaming") {
-    if (ciInfo.isCI) {
-      program.observability = getDefault("observability");
-    } else {
-      const { observability } = await prompts(
-        {
-          type: "select",
-          name: "observability",
-          message: "Would you like to set up observability?",
-          choices: [
-            { title: "No", value: "none" },
-            ...(program.framework === "fastapi"
-              ? [{ title: "LlamaTrace", value: "llamatrace" }]
-              : []),
-            { title: "Traceloop", value: "traceloop" },
-          ],
-          initial: 0,
-        },
-        questionHandlers,
-      );
+    const { observability } = await prompts(
+      {
+        type: "select",
+        name: "observability",
+        message: "Would you like to set up observability?",
+        choices: [
+          { title: "No", value: "none" },
+          ...(program.framework === "fastapi"
+            ? [{ title: "LlamaTrace", value: "llamatrace" }]
+            : []),
+          { title: "Traceloop", value: "traceloop" },
+        ],
+        initial: 0,
+      },
+      questionHandlers,
+    );
 
-      program.observability = observability;
-    }
+    program.observability = observability;
   }
 
   if (!program.modelConfig) {
@@ -230,132 +178,128 @@ export const askProQuestions = async (program: QuestionArgs) => {
   }
 
   if (!program.dataSources) {
-    if (ciInfo.isCI) {
-      program.dataSources = getDefault("dataSources");
-    } else {
-      program.dataSources = [];
-      // continue asking user for data sources if none are initially provided
-      while (true) {
-        const firstQuestion = program.dataSources.length === 0;
-        const choices = getDataSourceChoices(
-          program.framework,
-          program.dataSources,
-          program.template,
-        );
-        if (choices.length === 0) break;
-        const { selectedSource } = await prompts(
-          {
-            type: "select",
-            name: "selectedSource",
-            message: firstQuestion
-              ? "Which data source would you like to use?"
-              : "Would you like to add another data source?",
-            choices,
-            initial: firstQuestion ? 1 : 0,
-          },
-          questionHandlers,
-        );
+    program.dataSources = [];
+    // continue asking user for data sources if none are initially provided
+    while (true) {
+      const firstQuestion = program.dataSources.length === 0;
+      const choices = getDataSourceChoices(
+        program.framework,
+        program.dataSources,
+        program.template,
+      );
+      if (choices.length === 0) break;
+      const { selectedSource } = await prompts(
+        {
+          type: "select",
+          name: "selectedSource",
+          message: firstQuestion
+            ? "Which data source would you like to use?"
+            : "Would you like to add another data source?",
+          choices,
+          initial: firstQuestion ? 1 : 0,
+        },
+        questionHandlers,
+      );
 
-        if (selectedSource === "no" || selectedSource === "none") {
-          // user doesn't want another data source or any data source
+      if (selectedSource === "no" || selectedSource === "none") {
+        // user doesn't want another data source or any data source
+        break;
+      }
+      switch (selectedSource) {
+        case "exampleFile": {
+          program.dataSources.push(EXAMPLE_FILE);
           break;
         }
-        switch (selectedSource) {
-          case "exampleFile": {
-            program.dataSources.push(EXAMPLE_FILE);
-            break;
-          }
-          case "file":
-          case "folder": {
-            const selectedPaths = await selectLocalContextData(selectedSource);
-            for (const p of selectedPaths) {
-              program.dataSources.push({
-                type: "file",
-                config: {
-                  path: p,
-                },
-              });
-            }
-            break;
-          }
-          case "web": {
-            const { baseUrl } = await prompts(
-              {
-                type: "text",
-                name: "baseUrl",
-                message: "Please provide base URL of the website: ",
-                initial: "https://www.llamaindex.ai",
-                validate: (value: string) => {
-                  if (!value.includes("://")) {
-                    value = `https://${value}`;
-                  }
-                  const urlObj = new URL(value);
-                  if (
-                    urlObj.protocol !== "https:" &&
-                    urlObj.protocol !== "http:"
-                  ) {
-                    return `URL=${value} has invalid protocol, only allow http or https`;
-                  }
-                  return true;
-                },
-              },
-              questionHandlers,
-            );
-
+        case "file":
+        case "folder": {
+          const selectedPaths = await selectLocalContextData(selectedSource);
+          for (const p of selectedPaths) {
             program.dataSources.push({
-              type: "web",
+              type: "file",
               config: {
-                baseUrl,
-                prefix: baseUrl,
-                depth: 1,
+                path: p,
               },
             });
-            break;
           }
-          case "db": {
-            const dbPrompts: prompts.PromptObject<string>[] = [
-              {
-                type: "text",
-                name: "uri",
-                message:
-                  "Please enter the connection string (URI) for the database.",
-                initial: "mysql+pymysql://user:pass@localhost:3306/mydb",
-                validate: (value: string) => {
-                  if (!value) {
-                    return "Please provide a valid connection string";
-                  } else if (
-                    !(
-                      value.startsWith("mysql+pymysql://") ||
-                      value.startsWith("postgresql+psycopg://")
-                    )
-                  ) {
-                    return "The connection string must start with 'mysql+pymysql://' for MySQL or 'postgresql+psycopg://' for PostgreSQL";
-                  }
-                  return true;
-                },
+          break;
+        }
+        case "web": {
+          const { baseUrl } = await prompts(
+            {
+              type: "text",
+              name: "baseUrl",
+              message: "Please provide base URL of the website: ",
+              initial: "https://www.llamaindex.ai",
+              validate: (value: string) => {
+                if (!value.includes("://")) {
+                  value = `https://${value}`;
+                }
+                const urlObj = new URL(value);
+                if (
+                  urlObj.protocol !== "https:" &&
+                  urlObj.protocol !== "http:"
+                ) {
+                  return `URL=${value} has invalid protocol, only allow http or https`;
+                }
+                return true;
               },
-              // Only ask for a query, user can provide more complex queries in the config file later
-              {
-                type: (prev) => (prev ? "text" : null),
-                name: "queries",
-                message: "Please enter the SQL query to fetch data:",
-                initial: "SELECT * FROM mytable",
+            },
+            questionHandlers,
+          );
+
+          program.dataSources.push({
+            type: "web",
+            config: {
+              baseUrl,
+              prefix: baseUrl,
+              depth: 1,
+            },
+          });
+          break;
+        }
+        case "db": {
+          const dbPrompts: prompts.PromptObject<string>[] = [
+            {
+              type: "text",
+              name: "uri",
+              message:
+                "Please enter the connection string (URI) for the database.",
+              initial: "mysql+pymysql://user:pass@localhost:3306/mydb",
+              validate: (value: string) => {
+                if (!value) {
+                  return "Please provide a valid connection string";
+                } else if (
+                  !(
+                    value.startsWith("mysql+pymysql://") ||
+                    value.startsWith("postgresql+psycopg://")
+                  )
+                ) {
+                  return "The connection string must start with 'mysql+pymysql://' for MySQL or 'postgresql+psycopg://' for PostgreSQL";
+                }
+                return true;
               },
-            ];
-            program.dataSources.push({
-              type: "db",
-              config: await prompts(dbPrompts, questionHandlers),
-            });
-            break;
-          }
-          case "llamacloud": {
-            program.dataSources.push({
-              type: "llamacloud",
-              config: {},
-            });
-            program.dataSources.push(EXAMPLE_FILE);
-            break;
-          }
+            },
+            // Only ask for a query, user can provide more complex queries in the config file later
+            {
+              type: (prev) => (prev ? "text" : null),
+              name: "queries",
+              message: "Please enter the SQL query to fetch data:",
+              initial: "SELECT * FROM mytable",
+            },
+          ];
+          program.dataSources.push({
+            type: "db",
+            config: await prompts(dbPrompts, questionHandlers),
+          });
+          break;
+        }
+        case "llamacloud": {
+          program.dataSources.push({
+            type: "llamacloud",
+            config: {},
+          });
+          program.dataSources.push(EXAMPLE_FILE);
+          break;
         }
       }
     }
@@ -377,23 +321,19 @@ export const askProQuestions = async (program: QuestionArgs) => {
     ) {
       // if already set useLlamaParse, don't ask again
       if (program.dataSources.some((ds) => ds.type === "file")) {
-        if (ciInfo.isCI) {
-          program.useLlamaParse = getDefault("useLlamaParse");
-        } else {
-          const { useLlamaParse } = await prompts(
-            {
-              type: "toggle",
-              name: "useLlamaParse",
-              message:
-                "Would you like to use LlamaParse (improved parser for RAG - requires API key)?",
-              initial: false,
-              active: "Yes",
-              inactive: "No",
-            },
-            questionHandlers,
-          );
-          program.useLlamaParse = useLlamaParse;
-        }
+        const { useLlamaParse } = await prompts(
+          {
+            type: "toggle",
+            name: "useLlamaParse",
+            message:
+              "Would you like to use LlamaParse (improved parser for RAG - requires API key)?",
+            initial: false,
+            active: "Yes",
+            inactive: "No",
+          },
+          questionHandlers,
+        );
+        program.useLlamaParse = useLlamaParse;
       }
     }
   }
@@ -402,22 +342,17 @@ export const askProQuestions = async (program: QuestionArgs) => {
   if (isUsingLlamaCloud || program.useLlamaParse) {
     if (!program.llamaCloudKey) {
       // if already set, don't ask again
-      if (ciInfo.isCI) {
-        program.llamaCloudKey = getDefault("llamaCloudKey");
-      } else {
-        // Ask for LlamaCloud API key
-        const { llamaCloudKey } = await prompts(
-          {
-            type: "text",
-            name: "llamaCloudKey",
-            message:
-              "Please provide your LlamaCloud API key (leave blank to skip):",
-          },
-          questionHandlers,
-        );
-        program.llamaCloudKey =
-          llamaCloudKey || process.env.LLAMA_CLOUD_API_KEY;
-      }
+      // Ask for LlamaCloud API key
+      const { llamaCloudKey } = await prompts(
+        {
+          type: "text",
+          name: "llamaCloudKey",
+          message:
+            "Please provide your LlamaCloud API key (leave blank to skip):",
+        },
+        questionHandlers,
+      );
+      program.llamaCloudKey = llamaCloudKey || process.env.LLAMA_CLOUD_API_KEY;
     }
   }
 
@@ -426,50 +361,42 @@ export const askProQuestions = async (program: QuestionArgs) => {
     const vectorDb = "llamacloud";
     program.vectorDb = vectorDb;
   } else if (program.dataSources.length > 0 && !program.vectorDb) {
-    if (ciInfo.isCI) {
-      program.vectorDb = getDefault("vectorDb");
-    } else {
-      const { vectorDb } = await prompts(
-        {
-          type: "select",
-          name: "vectorDb",
-          message: "Would you like to use a vector database?",
-          choices: getVectorDbChoices(program.framework),
-          initial: 0,
-        },
-        questionHandlers,
-      );
-      program.vectorDb = vectorDb;
-    }
+    const { vectorDb } = await prompts(
+      {
+        type: "select",
+        name: "vectorDb",
+        message: "Would you like to use a vector database?",
+        choices: getVectorDbChoices(program.framework),
+        initial: 0,
+      },
+      questionHandlers,
+    );
+    program.vectorDb = vectorDb;
   }
 
   if (
     !program.tools &&
     (program.template === "streaming" || program.template === "multiagent")
   ) {
-    if (ciInfo.isCI) {
-      program.tools = getDefault("tools");
-    } else {
-      const options = supportedTools.filter((t) =>
-        t.supportedFrameworks?.includes(program.framework),
-      );
-      const toolChoices = options.map((tool) => ({
-        title: `${tool.display}${toolRequiresConfig(tool) ? " (needs configuration)" : ""}`,
-        value: tool.name,
-      }));
-      const { toolsName } = await prompts({
-        type: "multiselect",
-        name: "toolsName",
-        message:
-          "Would you like to build an agent using tools? If so, select the tools here, otherwise just press enter",
-        choices: toolChoices,
-      });
-      const tools = toolsName?.map((tool: string) =>
-        supportedTools.find((t) => t.name === tool),
-      );
-      program.tools = tools;
-    }
+    const options = supportedTools.filter((t) =>
+      t.supportedFrameworks?.includes(program.framework),
+    );
+    const toolChoices = options.map((tool) => ({
+      title: `${tool.display}${toolRequiresConfig(tool) ? " (needs configuration)" : ""}`,
+      value: tool.name,
+    }));
+    const { toolsName } = await prompts({
+      type: "multiselect",
+      name: "toolsName",
+      message:
+        "Would you like to build an agent using tools? If so, select the tools here, otherwise just press enter",
+      choices: toolChoices,
+    });
+    const tools = toolsName?.map((tool: string) =>
+      supportedTools.find((t) => t.name === tool),
+    );
+    program.tools = tools;
   }
 
-  await checkAskPostInstallAction();
+  program.postInstallAction = await askPostInstallAction(program);
 };
